requirements.txt
black==20.8b1
boto3==1.7.42
coverage==5.1
django-anymail==3.0
django-cors-headers==3.2.1
django-extra-fields==2.0.3
django-heroku==0.3.1
django-markdownx==3.0.1
django-nose==1.4.6
django-reversion==3.0.7
django-storages==1.9.1
Django==3.0.7
djangorestframework-gis==0.15
djangorestframework-simplejwt==4.3.0
djangorestframework==3.10.3
djoser==2.0.3
flake8==3.7.8
gunicorn==19.7.0
ipython==6.2.1
mailjet-rest==1.3.3
numpy==1.17.4
pinax-notifications==6.0.0
pinax-templates==2.0.6
psycopg2==2.7.3.2
pytz==2020.1
requests==2.20.0
roman==3.2